The part 2 to this problem
The password policy has changed and now the given character must be in exactly one of the given index positions. If it's in both the index position's, then it is invalid. If it isnt in any of the index position's, it is invalid. The program will ignore any other occurences of the character if it is not in the given index positions.