Dodge the trees.
This problem has a given input of trees that you have to count each tree that you hit with the movements of right 3 down 1. When you reach the end of the line instead of multiplying the data and taking up heaps of space, you mod the column position by the row length to wrap to the start of the next row.